(Almost) always use CPU Profiling 2.0 when profiling is enabled
Prior to the "no signals" workaround existing, we were using the legacy
profiler as a fallback in cases where the CPU Profiling 2.0 profiler
could/would cause issues.

Since the source of those issues is always related to the signal
sending/receiving mechanisms, we can now stop using the legacy profiler
as a fallback, and instead use the "no signals" workaround to replace it.

The only case where the legacy profiler is used is when it is force
enabled; in all other cases the CPU Profiling 2.0 profiler is used,
with or without the "no signals" workaround (as needed/configured).