Remote configuration handle Client SyncError
Whenever a Client::SyncError is raised when calling `client.sync`
the remote component rescues it, logs the sync error information
and continues with the execution.

The remote configuration backend expects to maintain the state
between requests even if there is a SyncError.

In the case of an unhandled error, we are still
resting the client's state.