Enable Remote configuration by default
Even with remote configuration enabled by default,
on this version of remote configuration and appsec.
If AppSec is disabled, the remote configuration worker thread won't start.

Since in AppSec, we do not support the feature of remote activation.
There is no need to have a background thread opening HTTP connections to
communicate with the  agent to request remote configuration information.

This might change once we support ASM remote activation or other components,
such as tracer, start using remote configuration since there would
be remote configuration information to listen to.