Add support for Unix Domain Socket (UDS) configuration via `DD_TRACE_AGENT_URL`
**What does this PR do?**:

This PR extends the `AgentSettingsResolver` to allow the
`DD_TRACE_AGENT_URL` to contain a Unix Domain Socket (UDS) configuration
string, for instance `unix:///some/path/to/the/apm.socket`.

It also adds a warning if there's configuration for both http and
unix domain socket (and in that case it will prioritize the http
variant).

**Motivation**:

This was missing for feature parity with other Datadog client libraries,
and it was annoying a lot of customers in #1967 (thanks for the
patience, folks!).

**Additional Notes**:

The change is quite small because dd-trace-rb already supported using
unix domain sockets -- you just couldn't configure them via this
environment variable. Thus, the only thing that changed was parsing it
correctly.

**How to test the change?**:

Configure an agent with a uds socket and report data through it.

Fixes #1967