AppSec remote
Make sure to handle product updates properly.

There is no warranty that a product update will only contain a single
product key.

In the case of `ASM` product, the schema allows for multiple keys:
`rules_override`, and `exclusions`

In that case we should make sure that we properly extract
each product information orrectly. That way the RuleMerger can combine
all the information correctly.

Since we have modify the content that we pass to the RulerMerger
we had to modify a little its code.