[PROF-7622] Add support for profiling Ruby 3.3.0-preview1
**What does this PR do?**:

Ruby 3.3.0-preview1 is the first release of the 3.3 branch. It includes
a pretty big change: it drops MJIT (one of the two JIT implementations
being shipped, the other being YJIT).

This affects the profiler because we were (ab)using an internal header
that Ruby included only for supporting MJIT, and this header is no more.

The solution here is to use the same solution we were using for
Ruby < 2.6. Quoting the `NativeExtensionDesign.md` file:

> Because these private header files are not included in regular Ruby installations, we have two different workarounds:
>
> 1. for Ruby versions >= 2.6 we make use use the Ruby private MJIT header
> 2. for Ruby versions < 2.6 (legacy Rubies) we make use of the `debase-ruby_core_source` gem

Thus, the strategy going forward will be to also use the
`debase-ruby_core_source` gem to support Ruby 3.3+.

Furthermore, there were a few changes to the APIs exported by Ruby
around ractors, that I also had to fix.

**Motivation**:

Make sure we keep up-to-date with latest Rubies, and that we don't
block our, and our client's experiments with these versions.

**Additional Notes**:

Because we aren't yet running Ruby 3.3 on our CI, you'll have to
take my word for it that this change is correct. (Or check for
yourself locally).

Separately from this PR, I'll submit a PR to add the
3.3.0-preview1 headers to `debase-ruby_core_source`, as I've been
testing using a local fork.

Furthermore, I plan to bootstrap Ruby 3.3 testing in our CI so that
we can validate that support for 3.3 keeps working.

But I don't see any reason to delay this PR before those two others.

**How to test the change?**:

This change can only be tested locally at the moment: install
Ruby 3.3.0-preview1 in your favorite way, and then point your `Gemfile`
to use
<https://github.com/DataDog/debase-ruby_core_source/tree/datadog/add-ruby-3.3.0-preview1-headers>,
and you'll be able to run the profiling specs as usual.

Fixes #2845