Bootstrap Ruby 3.3 testing (with 3.3.0-preview1)
**What does this PR do?**:

This PR goes through every file that needs to be modified to support
a new Ruby version (usually helpfully marked with "ADD NEW RUBIES HERE")
and adds Ruby 3.3 there.

This allows us to start testing with Ruby 3.3 in CI, ahead of the
expected release in December.

**Motivation**:

Having Ruby 3.3 in CI allows us to spot issues early and to be ready
for it once it releases. (I've also heard of at least one downstream
gem that integrates with ddtrace and tests with the latest Ruby, and
we want to be in good shape for this kind of testing as well).

In particular, the profiler needs a few tweaks to work with Ruby 3.3
(see #2860) and this allows us to make sure those tweaks are working.

**Additional Notes**:

This PR sits atop #2860 and
<https://github.com/ruby-debug/debase-ruby_core_source/pull/8> so that
the profiler test suite can be run as well.

**How to test the change?**:

I've modified the usual testing docker images, so the usual testing
approach used for other Ruby versions documented in
`DevelopmentGuild.md` also applies to this new Ruby version.