Remove unused AppSec instrumentation options Non of our instrumention accept any options.
We are cargo culting a pattern from tracing
that is not used in AppSec.

If we ever need to provide instrumentation optinos for AppSec
we would be ablet to implement it or bring this code back again.

Also, exposing ENV variables that are not used for anything is a bad
pattern :(