Fix flaky profiler integration test due to thread naming race
**What does this PR do?**:

This PR tweaks the way we set the thread names in the profiler threads.

Previously, each worker thread set its own name; now, the caller thread
also sets the name.

**Motivation**:

Because Ruby threads cannot be named at creation (which is weird --
Java allows this), setting the name of a thread is always a race:
there's a period after the `Thread` object is created where the
thread will have no name.

In that period, calling `Thread.list` will not see the name.

I suspect this is the issue that led to a flaky test reported by
@tonycthsu
(<https://app.circleci.com/pipelines/github/DataDog/dd-trace-rb/10382/workflows/ed4867b5-aa41-4179-a27c-fdb73b8d86c2/jobs/385644>):
the integration tests check that all profiler threads are alive
by checking for their names, but the thread may not have yet set
its own name.

Setting the name twice is harmless, so I went for this approach of
setting it in both the parent as well as the created threads.

**Additional Notes**:

N/A

**How to test the change?**:

I was able to reproduce the "missing names" issue on a trivial Ruby
script that did

```ruby
require 'ddtrace'

Datadog::Profiling.start_if_enabled
puts Thread.list
```