[PROF-7645] Include state of "no signals" workaround when reporting profiles
**What does this PR do?**:

In #2873 we introduced an alternative mode to the profiler, called the
"no signals" workaround.

This PR makes use of the new libdatadog internal metadata reporting
support to send the state of this setting along with reported profiles.

**Motivation**:

We'll use this to track how many customers are using the "no signals"
workaround, as well as when debugging issues.

**Additional Notes**:

This PR will break in CI since it depends on the libdatadog support for
the internal metadata reporting, which has not been released yet as of
this writing (https://github.com/datadog/libdatadog/pull/181).

My plan is to open a separate PR to update the libdatadog version in use
by ddtrace, and then rebase this PR on top of it.

This PR has been tested locally, and it's working and ready for review.

Once it's approved, I'll make sure it's merged only after the new
libdatadog version is available and CI is green.

**How to test the change?**:

This change includes test coverage. Testing it locally requires building
libdatadog locally as well -- feel free to ping me and I can guide you
through it.