Make sure the client interceptor actually returns the response
Datadog::Tracing::Contrib::GRPC::DatadogInterceptor::Client#trace is
currently causing all GRPC client operations to return 0.0 instead of
whatever they were supposed to return. This is a bit of a Ruby gotcha.

This method returns :three

```ruby
def broken_method
    begin
        :one
    rescue => e
        :two
    else
        :three
    end
end
```

So when `#trace` is putting `yield` in a `begin` block with an `else`,
it's not actually returning the result of yield; it's returning the
result of the else block.

Fix it by explicitly returning the result of yield.