[PROF-7654] Bump libdatadog dependency to version 3
**What does this PR do?**:

This PR bumps the libdatadog dependency on dd-trace-rb to version 3
(specifically, '~> 3.0.0.1.0').

It also includes small changes to code/tests to adapt to changes in
libdatadog and get a green CI.

**Motivation**:

Libdatadog 3 includes the new "internal metadata" feature which we'll
use in #2927.

**Additional Notes**:

Despite the jump in major version, the changes between libdatadog 2
and 3 were minimal. The version was bumped because the library is
following semantic versioning, and a public API was changed.

**How to test the change?**:

Running the profiling specs exercises libdatadog, and thus we have
quite a lot of test coverage for it.