Record thread_invoke_location for each per_thread_context
The `thread_invoke_location` can be used as a fallback alternative for
threads that have no name, and is what Ruby shows in `Thread#to_s` as
well.

This commit only gathers the invoke location and stores it in the
context; this information is not yet propagated in the pprof, that
will come in a later commit.