Fix theoretical line number calculation bug in profiler for Ruby 3.2+
**What does this PR do?**:

This PR fixes a theoretical line number calculation bug in the profiler.

Why theoretical?

Because the `! pc` branch it's under is never called by the profiler
currently (the `ddtrace_rb_profile_frames` checks for a valid `pc` before
calling the line number calculation). Nevertheless, I think it's
worth fixing as in the future our code may change.

Since Ruby 3.2.0+ (<https://github.com/ruby/ruby/pull/6430>) the
`first_lineno` that Ruby stores for a set of instructions is stored
as a regular `int` and not as a Ruby `VALUE`.

The `calc_pos` function inside `private_vm_api_access.c` was copied
from the Ruby sources before that refactoring, and thus was still
trying to convert the Ruby `VALUE` to an `int`, which is not
correct for modern Rubies.

Since this code path is only for really rare situations -- In fact,
older versions of the VM code
(see https://github.com/ruby/ruby/commit/ea6062898ad0d66ede0a1866028c0605c357e2cb )
even had a comment saying this could happen "during VM bootup",
this was not caught on any of our testing.

TL;DR I discovered this when I was reading the code and working on
something else and got curious, and tested this by manually forcing
`pc = NULL` and then re-running our test suite.

**Motivation**:

Make sure in the future we don't get tripped by this potential bug.

**Additional Notes**:

N/A

**How to test the change?**:

As I mentioned above, this issue is for something we don't expect
can happen currently, deep inside `private_vm_api_access.c`.

Exposing this function for unit testing would be more work and add
more complexity for something that's not expected to trigger, so
I've opted to only test manually with the change exposed above.