[PROF-7440] Use invoke location as a fallback for nameless threads in the profiler
**What does this PR do?**:

Naming Ruby threads is optional, and a lot of Ruby code still creates
threads and doesn't name them.

As a workaround for Ruby threads that have no name, this PR makes the
profiler use the "thread invoke location", which is what Ruby shows in
the `to_s`/`inspect` of a thread, for instance:

```
[1] pry(main)> Thread.list.last.name
=> nil
[2] pry(main)> Thread.list.last.to_s
=> "#<Thread:0x00005934cc0143e8 thread-name.rb:3 sleep>"
```

The "invoke location" is the file and line where the block that
starts the thread is defined.

**Motivation**:

Having nameless threads is annoying because we show this information
in several places of our UX, and with the timeline feature this will be
even more visible.

**Additional Notes**:

Since Ruby doesn't have an API to get this information (other than
calling `to_s` and then parsing its output), and because we want to
avoid allocating any Ruby objects while the profiler is taking a sample,
I chose to add to `private_vm_api_access.c` an API to get this
information that is basically a heavily simplified version of
`Thread#to_s`.

**How to test the change?**:

This change includes code coverage. Also, profile any app that starts
and doesn't name threads, and you'll start seeing threads named
"some_file.rb:some_line" in the profiler.