When working on https://github.com/DataDog/dd-trace-rb/pull/2936 I noticed that calling `SafeDup.frozen_or_dup({a: :b})` on ruby 3.2 would break; because it tries to call `+(-){a: :b}` which doesn't work.
I had to ensure the `SafeDup` module worked for all objects.

I decided to extract that work into a separate PR to ease the work of reviewing.

The method `frozen_or_dup` and `frozen_dup` for ruby 2.3 and forward only expect to receive String values.
This could lead to future errors.
By checking if the value is a String we can warranty the same behaviour for all kinds of objects

Also, I added some extract refimients to be able to call `#dup` on `true, false, 1, 1.0` values.