Tweak versions of google-protobuf used for testing in CI
**What does this PR do?**:

This PR tweaks the versions of google-protobuf used for testing to
make sure to avoid the latest 3.24.0.rc.1 being a bit problematic for
some legacy Rubies.

This caused some failures in our CI:
https://app.circleci.com/pipelines/github/DataDog/dd-trace-rb/10871/workflows/90c509f6-a12a-441a-8e1d-ad0b1567efed/jobs/405709

And has been reported upstream:
https://github.com/protocolbuffers/protobuf/issues/13307

**Motivation**:

Avoid problematic google-protobuf versions in CI.

**Additional Notes**:

The google-protobuf gem has some history of having incorrect
packaging, which is why we already had a bunch of versions
excluded (if you look at the PR diffs).

One more for the pile!

Google-protobuf is a dependency needed for the old profiler bits,
and the profiler includes some help/detection to guide customers
around these packaging issues when they are detected.

Historically, not a lot of customers reached out with issues
regarding google-protobuf, so I think it's mostly a problem
for our CI which is more complex since it needs to run on a
bunch of different rubies/platforms/etc.

**How to test the change?**:

Validate that CI is green!