Enable profiler "no signals" workaround for passenger web server
**What does this PR do?**:

This PR adds one more case where we automatically need to apply the
profiler "no signals" workaround (added in #2873): when the Ruby app
is running inside the passenger web server.

**Motivation**:

This makes sure the profiler does not cause issues out-of-the-box
for customers running passenger.

**Additional Notes**:

Passenger detection was inspired on
https://stackoverflow.com/a/12246652/4432562 .

This was reported in https://github.com/datadog/dd-trace-rb/issues/2976
and I suspect at least one other customer has ran into this issue.

Separate from this PR, I plan to improve our documentation, as well as
try to get in touch with the passenger folks to get a fix upstream.
I'm tracking that work in the mentioned issue.

I'm also considering updating the integration apps to include passenger,
but I wanted to get in the fix asap, and will circle back to that one in
a separate PR if I can get it to work sanely.

**How to test the change?**:

Running the test-case submitted for #2976:
https://github.com/elliterate/ddtrace-passenger-example can be used
to validate that the profiler auto-enables the no signals workaround.