[PROF-7440] Add fallback name/invoke location for unnamed threads started in native code
**What does this PR do?**:

This PR adds a fallback invoke location (which gets used as a name) for
threads that a) Are unnamed, and b) Were started from native code.

The added invoke location is the fixed string
`(Unnamed thread from native gem)`.

**Motivation**:

We previously (in #2950) started using invoke location for threads
that do not have a name BUT when a thread is started from native
code it would have an empty name.

Having an empty name is not particularly helpful, so I decided to
instead have a simple placeholder that at least provides an hint on
what the thread may be.

**Additional Notes**:

We don't yet do any resolving of symbols from native code, but in
the future I guess we may try to see if the native code has a name
at the C level, and use that as a fallback, but that's much more
than a one-liner change :)

**How to test the change?**:

Change includes code coverage.