Remove unused Barrier#wait_next functionality
**What does this PR do?**:

This PR removes the `Barrier#wait_next` functionality which we ended
up not making use of.

**Motivation**:

In #2990 we ran into issues with its current specs, and since we're
not using this functionality, it seems like removing it for now
seems to be easier than investing more time to fix flakiness in the
specs.

**Additional Notes**:

N/A

**How to test the change?**:

Existing tests already cover the `Barrier` APIs.