Force-disable the no signals workaround for gitlab testing
Gitlab makes use of the rugged gem, which triggers the automatic
"no signals" workaround usage, as documented in
<https://docs.datadoghq.com/profiler/profiler_troubleshooting/ruby/#unexpected-failures-or-errors-from-ruby-gems-that-use-native-extensions-in-dd-trace-rb-1110>

But in practice the endpoints we test it aren't affected, so we prefer
to run the profiler in its default, better accuracy, mode.

Thus, let's disable the "no signals" workaround explicitly via
environment variable (note: this is not recommended in most cases!).