Remove profiler_submission.rb benchmark
**What does this PR do?**:

This PR removes the `profiler_submission.rb` benchmark.

**Motivation**:

This benchmark is no longer relevant, as it only tested the old Ruby
profiler, and not the new CPU Profiling 2.0 profiler.

The old profiler is no longer in use and will be removed soon, and I
spotted this benchmark failing in the benchmarking platform and rather
than fix it, I decided to remove it instead.

In the future, we may want to add a benchmark to replace it, but the
new profiler internals are different enough that it's non-trivial
to update the existing benchmark.

**Additional Notes**:

Update to benchmarking platform is at
https://github.com/DataDog/benchmarking-platform/pull/16 (Datadog
internal link).

**How to test the change?**:

Validate that CI is still green.