Fix warnings not including defaults when defaults were used
The `configured_hostname` and `configured_port` don't include the
defaults, so the resulting message was a bit confusing when defaults
were picked for them.

By using `hostname` and `port` instead, we make sure the warning
message shows the values that will be picked.

(I did notice a bit of a weird recursion where calling `hostname` and
`port` does call back into `mixed_http_and_uds?` via `should_use_uds?`,
but it's fine because we already set the `@mixed_http_and_uds` before
doing so. It is a bit awkward but... I feel in a way this is reflecting
the actual complexity of all these priorities and overrides and whatnot,
so I'll leave it to someone else to try to untangle a bit better in
the future.)