Correctly print uds path in warning message
`uds_path` vs `parsed_url` is similar to `hostname` vs
`configured_hostname`: The `uds_path` is the actual final value that's
going to be included in the `AgentSettingsResolver` output, so it
will be `nil` when `mixed_http_and_uds?` is true.

In contrast, `parsed_url` contains the configured value which caused
the configuration mismatch, and so that's the correct one to print.