Refactor `Flush` to receive `internal_metadata` without knowing what's in it
During review of #2927, @marcotc did point out that getting the
internal metadata information from the `Datadog::Profiling::Component`
to the `Flush` object was quite a lot of changes.

As I'm preparing to add a second entry to `internal_metadata`, I'm
refactoring this code to try to change it so that fewer components
need to be touched when we want to add more stuff.