Pass `internal_metadata` into `Exporter` as a hash
This is one more step in a refactor in preparation for adding more
internal metadata. By having the `Exporter` receive the
`internal_metadata` as a opaque hash it cares nothing about, we can
easily add more entries directly in the `Profiling::Component` without
needing to touch a lot of classes.