[PROF-7440] Report status of timeline feature as internal metadata
**What does this PR do?**:

This PR adds support for reporting the status of the timeline feature
together with profiles, via the internal metadata field.

I ended up doing a bunch of refactoring to make it easier to adds new
reported fields as internal metadata; it may be easier to review this
PR commit-by-commit.

**Motivation**:

While this feature is off-by-default, having this internal metadata
makes it easier to spot profiles that have/don't have this data
without actually opening them up in the UX.

**Additional Notes**:

N/A

**How to test the change?**:

This change includes test coverage. It's possible internally at
Datadog to see this information in profile listings (ping me if you'd
like to know more).