Introduce new `LinuxTidFallback` class
This class is the bridge between the profiler and the C-level
implementation for `linux_tid_from_pthread`.

It is some amount of boilerplate vs just calling
`linux_tid_from_pthread` from C, but I like that it makes the
feature easy to disable as well as easy to test, without adding
more complexity to the `ThreadContext` collector.