Disable profiler on Ruby 3.3 due to incompatibility
**What does this PR do?**:

In #3053 I reported the discovery that the profiler is currently not
working on Ruby 3.3-head (reported by a customer that tests with
head in CI).

A full fix for #3053 seems like it may take a while, so until it's
available, let's gracefully disable the profiler in Ruby 3.3, so that
instead of a compilation failure, customers just get a nice message
saying we don't support 3.3 right now.

**Motivation**:

Improve experience for current customers testing with 3.3-head and
future customers trying to profile 3.3.

**Additional Notes**:

Customers will be shown the following message when trying to profile
on Ruby 3.3:

```
WARN -- ddtrace: [ddtrace] Profiling was requested but is not supported,
profiling disabled: Your ddtrace installation is missing support for the
Continuous Profiler because the profiler in the current version of
ddtrace does not yet support Ruby version 3.3. (See
https://github.com/datadog/dd-trace-rb/issues/3053 for details). Try
upgrading to the latest ddtrace, as this issue may have been fixed by
now. For help solving this issue, please contact Datadog support at
<https://docs.datadoghq.com/help/>. You can also check out the
Continuous Profiler troubleshooting page at
<https://dtdg.co/ruby-profiler-troubleshooting>.
```

**How to test the change?**:

Try to profile ruby-3.3 (any version) and confirm the profiler
disables itself gracefully.