Move `can_use_process_vm_readv` inside `linux_tid_from_pthread`
It occured to me that it was kinda overkill to expose `read_safely`
outside of the `linux_tid_from_pthread` module if the intention was
just to use it to check if `process_vm_readv` was working.

Thus, I folded that check inside that file, and only expose the
functionality we need -- a bool telling us if it worked or not.