[PROF-7307] Include 'ruby vm type' in profiler allocation samples
**What does this PR do?**:

This PR builds extends the existing support for the `ThreadContext`
collector to sample allocations to also record the "ruby vm type" of
sampled objects.

This "ruby vm type" is what Ruby internally represents in
its `enum ruby_value_type`, see
https://github.com/ruby/ruby/blob/84a12d657848dfb54f8cc556d344f017a793e95a/include/ruby/internal/value_type.h#L112

**Motivation**:

I'm working on tagging stacks of sampled objects with both the
object class, as well as the ruby vm type.

I'm not yet sure how we may expose the ruby vm type in the UX,
but having it is useful for development and debugging, and I
can imagine it being useful in advanced use cases
(e.g. a customer downloading a pprof to look at this data).

As libdatadog interns labels
(https://github.com/DataDog/libdatadog/pull/205), the cost
of tracking this data is quite low.

**Additional Notes**:

Allocation profiling is still not something that can be enabled;
e.g. right now we can only exercise it directly by calling the
`ThreadContext` collector.

**How to test the change?**:

Change includes test coverage.