Introduce explicit negative define for detecting T_MOVED
The previous `#ifdef RUBY_T_MOVED` was actually failing silently; and
because it's hard to test this helper, it was easy to miss since it
just meant a not-often-used element of the enum was omitted on all
Rubies.

By instead flipping the check to be not defined, we get a better
behavior: the default is to include the enum member, and if we get
the logic to disable incorrect, we get a compile error, not a silent
bug.