Refactor: Separate Profiling ThreadContext initialization from CpuAndWallTimeWorker
**What does this PR do?**

This PR separates the initialization of the
`Profiling::Collectors::ThreadContext` component from the initialization
for the `Profiling::Collectors::CpuAndWallTimeWorker`.

Now, both are created in `Datadog::Profiling::Component`.

**Motivation:**

Historically, these two classes have been moving apart, as the
profiler development evolves.

As more settings need to get added to the `ThreadContext`, it became
very awkward to keep changing the `CpuAndWallTimeWorker` to forward
more and more things.

**Additional Notes:**

N/A

**How to test the change?**

The existing test coverage already covers this.