Make recording of allocation types optional
I suspect the allocation types may get a bit expensive on legacy Rubies,
and plan to experiment more about this at some point.

For now, to enable quick experiments, I've decided to add a flag to make
it easy to toggle this behavior. I haven't decided if we'll ever expose
this flag to customers or not.

If you're reading this in a few months' time, and this disabling
feature was never used, go ahead and delete it.