Enable type checking for `Profiling::Collectors::CpuAndWallTimeWorker`
**What does this PR do?**

This PR adds the type signatures and enables type checking for the
`Profiling::Collectors::CpuAndWallTimeWorker` class.

**Motivation:**

I've extracted this from another branch I was working on, as it can
be cleanly merged right away.

**Additional Notes:**

This PR is on top of #3094 to avoid any conflicts, since that PR
changed some of the arguments to this class.

**How to test the change?**

Validate that type checking still passes!