Add pretty name for T_IMEMO VM objects
These show up from time to time and will be user-visible, so it's nice
to have a name for them.

We could consider hiding them, but my thinking is: what if there's
something in the user's code that causes a lot of them to be created?

That seems relevant to report on, even if they are VM objects, since
the user will probably want to change their code to avoid this.