Fix missing endpoint profiling when request_queuing is enabled in rack instrumentation
**What does this PR do?**

This PR tweaks the logic that decides if the profiler should collect
the resource from an active trace to also consider the span type
`proxy`, which is used when the `request_queuing` feature is in use
in the rack instrumentation.

Without this fix, the endpoint profiling feature (e.g. allowing
customers to drill into their profiles by endpoint) does not work
in combination with `request_queuing`, and the feature is hidden
in the profiler UX.

**Motivation:**

Make sure that customers that enable `request_queuing` are still able
to look at their profiles per-endpoint.

**Additional Notes:**

N/A

**How to test the change?**

The change includes test coverage. If you're interested in testing
it out with a real http app, you can do so by enabling the
`request_queuing` feature:

```ruby
Datadog.configure do |c|
  c.tracing.instrument :rack, request_queuing: true
end
```

and then perform a request to the app including the
`X-Request-Start` header:

```
$ curl --header "X-Request-Start: t=1693901946000" http://localhost:8080/foo
```

This will generate a `queue` span, and with this change, the endpoint
will show up in the right sidebar when looking at profiles for this
application. Without this change, it would not show up at all for
requests with the `queue` span.