Refactor: Simplify `Profiler` instance
**What does this PR do?**

This PR simplifies the `Profiler` class to receive only a single
`worker`, instead of a `collectors` array.

**Motivation:**

The `collectors` array was added a long time ago when we expected that
we may add more `collectors` in the future.

We never ended up making use of this extra flexibility, so I've
decided to simplify the code to remove this feature.

**Additional Notes:**

N/A

**How to test the change?**

This change includes code coverage. Since this is the core of
starting the profiler, it's pretty visible when it doesn't work ;)